Rename inverse.py to inverse_protocol.py to avoid 'Module not callable' error in mypy Generalize to arbitrary exponent
