+++ cirq/ops/moment.py
@@ -140,15 +140,13 @@ class Moment:
     def __pow__(self, power):
         if power == 1:
             return self
-        if power == -1:
-            inv = [
-                protocols.inverse(op, default=None) for op in self.operations
-            ]
-            for op in inv:
-                if not isinstance(op, raw_types.Operation):
-                    return NotImplemented
-            return Moment(inv)
-        return NotImplemented
+        new_ops = []
+        for op in self.operations:
+            new_op = protocols.pow(op, power, default=None)
+            if new_op is None:
+                return NotImplemented
+            new_ops.append(new_op)
+        return Moment(new_ops)
 
     def __len__(self):
         return len(self.operations)
