+++ cirq/protocols/inverse_protocol.py
