+++ cirq/protocols/inverse_protocol_test.py
